apps/api/pyproject.toml
[project]
name = "quanta-copilot-api"
version = "0.1.0"
description = "Source-Grounded Research Copilot API"
authors = [
    {name = "Quanta Copilot Team"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "sqlmodel>=0.0.14",
    "requests>=2.31.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "tiktoken>=0.5.0",
    "markdown>=3.5.0",
    "trafilatura>=2.0.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "PyMuPDF>=1.23.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
]

[project.optional-dependencies]
# Optional heavy dependencies
ocr = [
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
]

playwright = [
    "playwright>=1.40.0",
]

qdrant = [
    "qdrant-client>=1.7.0",
]

celery = [
    "celery>=5.3.0",
    "redis>=5.0.0",
]

# Development dependencies
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88
